Break up/rework TypeScript test suite
