+++ test/integration/typescript-hmr/pages/hello.tsx
@@ -1,0 +1,1 @@
+export default () => <p>Hello World</p>
