+++ test/integration/typescript-hmr/pages/type-error-recover.tsx
