+++ test/integration/typescript/test/index.test.js
@@ -1,26 +1,56 @@
 /* eslint-env jest */
 /* global jasmine */
 import { join } from 'path'
-import { renderViaHTTP, findPort, launchApp, killApp } from 'next-test-utils'
+import cheerio from 'cheerio'
+import {
+  renderViaHTTP,
+  nextBuild,
+  findPort,
+  launchApp,
+  killApp,
+  File
+} from 'next-test-utils'
 
-import hmr from './hmr'
-import typescript from './typescript'
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 1000 * 60 * 2
 
-const context = {}
-jasmine.DEFAULT_TIMEOUT_INTERVAL = 1000 * 60 * 5
+const appDir = join(__dirname, '..')
+let appPort
+let app
+
+async function get$ (path, query) {
+  const html = await renderViaHTTP(appPort, path, query)
+  return cheerio.load(html)
+}
 
 describe('TypeScript Features', () => {
-  beforeAll(async () => {
-    context.appPort = await findPort()
-    context.appDir = join(__dirname, '../')
-    context.server = await launchApp(context.appDir, context.appPort)
+  describe('default behavior', () => {
+    beforeAll(async () => {
+      appPort = await findPort()
+      app = await launchApp(appDir, appPort)
+    })
+    afterAll(() => killApp(app))
 
-    // pre-build all pages at the start
-    await Promise.all([renderViaHTTP(context.appPort, '/hello')])
-    await Promise.all([renderViaHTTP(context.appPort, '/type-error-recover')])
+    it('should render the page', async () => {
+      const $ = await get$('/hello')
+      expect($('body').text()).toMatch(/Hello World/)
+    })
   })
-  afterAll(() => killApp(context.server))
 
-  hmr(context, (p, q) => renderViaHTTP(context.appPort, p, q))
-  typescript(context, (p, q) => renderViaHTTP(context.appPort, p, q))
+  it('should compile the app', async () => {
+    const output = await nextBuild(appDir, [], { stdout: true })
+    expect(output.stdout).toMatch(/Compiled successfully/)
+  })
+
+  describe('should compile with different types', () => {
+    it('should compile async getInitialProps for _error', async () => {
+      const errorPage = new File(join(appDir, 'pages/_error.tsx'))
+      try {
+        errorPage.replace('static ', 'static async ')
+        const output = await nextBuild(appDir, [], { stdout: true })
+        expect(output.stdout).toMatch(/Compiled successfully/)
+      } finally {
+        errorPage.restore()
+      }
+    })
+  })
 })
